+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## master
 
 * [bugfix] Capture password without echo only if terminal is present
+* [feature] Support for `shelly ssh` command
 
 ## 0.4.28 / 2014-03-25